LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/Create-wrong-scheme.any-expected.txt

CONSOLE MESSAGE: Wrong url scheme for WebSocket http://web-platform.test:49001/echo

PASS Create WebSocket - Pass a URL with a non ws/wss scheme - SYNTAX_ERR is thrown
